requirements.txt
# PolicyGenerator-AI Requirements
# Author: Ayi NEDJIMI

python>=3.11
openai>=1.3.0
python-docx>=1.1.0
reportlab>=4.0.7
jinja2>=3.1.2
pyyaml>=6.0.1
structlog>=23.2.0
fastapi>=0.104.1
pytest>=7.4.3
